+++ CHANGELOG.md
@@ -2,6 +2,7 @@ CHANGELOG
 =========
 
 - Use the default config file as a fallback if options are missing in user config file
+- Do not show previously shown notifications
 
 v0.1.1
 ------